Module 8 Portfolio Project
Write a program that creates a Person class that contains string that represent the first and last name of a person and their age. You will need to create a Queue class that will store each person in the queue and can sort the queue based on last name or age.

Prompt the user of the program to add five people to the queue. Your program should provide the contents of the queue and then sort the queue using the quick sort in two ways:

Descending order by last name
Descending order by age